refactor(framework): Restructure DP-related components in flwr.common
