+++ framework/py/flwr/client/dpfedavg_numpy_client.py
@@ -21,8 +21,8 @@ import numpy as np
 
 from flwr.client.numpy_client import NumPyClient
 from flwr.common import Config, NDArrays, Scalar
-from flwr.common.dp import add_gaussian_noise, clip_by_l2
 from flwr.common.logger import warn_deprecated_feature
+from flwr.compat.common.dp import add_gaussian_noise, clip_by_l2
 
 
 class DPFedAvgNumPyClient(NumPyClient):
