+++ framework/py/flwr/clientapp/mod/centraldp_mods.py
@@ -33,10 +33,11 @@ from flwr.clientapp.typing import ClientAppCallable
 from flwr.common import log
 from flwr.common.constant import ErrorCode
 from flwr.common.differential_privacy import (
+    KEY_CLIPPING_NORM,
+    KEY_NORM_BIT,
     compute_adaptive_clip_model_update,
     compute_clip_model_update,
 )
-from flwr.common.differential_privacy_constants import KEY_CLIPPING_NORM, KEY_NORM_BIT
 
 
 # pylint: disable=too-many-return-statements
