+++ framework/py/flwr/clientapp/mod/dp_mods_test.py
@@ -28,7 +28,7 @@ from flwr.app import (
 )
 from flwr.app.message_type import MessageType
 from flwr.common import NDArray
-from flwr.common.differential_privacy_constants import KEY_CLIPPING_NORM, KEY_NORM_BIT
+from flwr.common.differential_privacy import KEY_CLIPPING_NORM, KEY_NORM_BIT
 
 from .centraldp_mods import adaptiveclipping_mod, fixedclipping_mod
 from .localdp_mod import LocalDpMod
