+++ framework/py/flwr/serverapp/strategy/dp_adaptive_clipping.py
@@ -27,12 +27,13 @@ import numpy as np
 from flwr.app import Array, ArrayRecord, ConfigRecord, Message, MetricRecord
 from flwr.common import log
 from flwr.common.differential_privacy import (
+    KEY_CLIPPING_NORM,
+    KEY_NORM_BIT,
     adaptive_clip_inputs_inplace,
     add_gaussian_noise_inplace,
     compute_adaptive_noise_params,
     compute_stdv,
 )
-from flwr.common.differential_privacy_constants import KEY_CLIPPING_NORM, KEY_NORM_BIT
 
 from ..exception import AggregationError
 from ..grid import Grid
