+++ framework/py/flwr/serverapp/strategy/dp_fixed_clipping.py
@@ -27,13 +27,11 @@ import numpy as np
 from flwr.app import Array, ArrayRecord, ConfigRecord, Message, MetricRecord
 from flwr.common import log
 from flwr.common.differential_privacy import (
+    CLIENTS_DISCREPANCY_WARNING,
+    KEY_CLIPPING_NORM,
     add_gaussian_noise_inplace,
     compute_clip_model_update,
     compute_stdv,
-)
-from flwr.common.differential_privacy_constants import (
-    CLIENTS_DISCREPANCY_WARNING,
-    KEY_CLIPPING_NORM,
 )
 
 from ..grid import Grid
